requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
blis==0.7.10
catalogue==2.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
confection==0.1.3
cymem==2.0.8
Django==4.2.5
djangorestframework==3.14.0
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.6.0/en_core_web_lg-3.6.0-py3-none-any.whl#sha256=9012584a8fcf5bc61bedb13a4b511fb81ffff9f770958ee06366671be10632e9
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.6.0/en_core_web_md-3.6.0-py3-none-any.whl#sha256=1618fc62e7aedc2b778caa3ea6e9c8280beac3b83fe7b93316cbb9e8f3abe7c5
idna==3.4
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.3
murmurhash==1.0.10
numpy==1.26.0
packaging==23.1
pathy==0.10.2
preshed==3.0.9
pydantic==2.3.0
pydantic_core==2.6.3
pyotp==2.9.0
pytz==2023.3.post1
requests==2.31.0
smart-open==6.4.0
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.7
thinc==8.1.12
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
urllib3==2.0.4
wasabi==1.1.2
